[mod/#33] 지민 언니 코리 반영 4
