+++ app/src/main/java/com/haphap/app/data/model/register/RegisterDropDownItemModel.kt
@@ -1,0 +1,6 @@
+package com.haphap.app.data.model.register
+
+data class RegisterDropDownItemData(
+    val id: String,
+    val text: String,
+)
