+++ app/src/main/java/com/haphap/app/presentation/register/type/RegisterPassResultType.kt
@@ -1,29 +1,7 @@
-package com.haphap.app.presentation.register
+package com.haphap.app.presentation.register.type
 
 import androidx.annotation.DrawableRes
-import androidx.compose.ui.graphics.Color
 import com.haphap.app.R
-import com.haphap.app.core.designsystem.theme.HapHapColors
-
-enum class DropDownItemSelectionState {
-    SELECTED,
-    UNSELECTED,
-}
-data class RegisterDropDownItemStyle(
-    val backgroundColor: Color,
-    val textColor: Color,
-)
-
-fun DropDownItemSelectionState.toStyle(colors: HapHapColors): RegisterDropDownItemStyle = when (this) {
-    DropDownItemSelectionState.SELECTED -> RegisterDropDownItemStyle(
-        backgroundColor = colors.sub100,
-        textColor = colors.primary500,
-    )
-    DropDownItemSelectionState.UNSELECTED -> RegisterDropDownItemStyle(
-        backgroundColor = colors.gray50,
-        textColor = colors.gray500,
-    )
-}
 
 enum class PassResultStatus(val text: String) {
     PASS("합격했어요"),
